Relocate utility functions into their own files
